Added the inverse of the covariance operator. The covariance is NOT YET RESCALED in this commit. This has the additional capabilities (see option -C) but works identically to the previous version. I am currently debugging the effect of the rescaled covariance on the inversion, before committing to the SVN.
Usage: -C  [FWD|INV] rFile_Model wFile_Model [rFile_Cov rFile_Prior]

 -C FWD rFile_Model wFile_Model [rFile_Cov rFile_Prior]
  Applies the model covariance to produce a smooth model output
  Optionally, also specify the prior model to compute resistivities
  from model perturbation: m = C_m^{1/2} \tilde{m} + m_0

 -C INV rFile_Model wFile_Model [rFile_Cov rFile_Prior]
  Applies the inverse of model covariance (if implemented)
  Optionally, also specify the prior model to compute starting
  perturbation for the inversion: \tilde{m} = C_m^{-1/2} (m - m_0)
  WARNING: may be poorly conditioned if the smoothing is too strong;
  always check your model for white noise after using this option.

NOTE: If you wish to use the inverse of the covariance to restart your inversion
with a different prior or covariance parameters, make sure that your starting
model does not contain significant white noise!!!!
The covariance is a poorly conditioned operator, essentially multiplying the
small scale features by zero; its inverse is therefore not well defined.
For example, if you apply alpha=0.1 covariance over 8 times, or alpha=0.3
covariance over 3 times, multiplication by the inverse covariance will produce
white noise of a large magnitude (3-4 orders of magnitude, comparable to that of
our models). It is safe to apply a strong covariance, but not to apply its inverse.
Until we find a way to regularize this, I am not making this part of the inversion,
but rather a separate option that may be used with care.