MAJOR CHANGES. MAY NOT WORK AS EXPECTED.
First, I rescaled the covariance, so that we no longer get unreasonable conductivity
variations when the covariance is applied several times. The covariance file means
the same thing as before, but the resultant model perturbations now have different magnitude
(the "rough" model has bigger magnitude variations than the "smooth" model).

With the new covariance, the default damping parameter and line search step had to
be adjusted. I used this opportunity to change the penalty functional as follows:

P(m) = (d-f(m))^T C_d^{-1} (d-f(m)) / N_d + lambda * (m-m_{prior})^T C_m^{-1} (m-m_{prior}) / N_m

where N_d is the number of real data values and N_m is the number of model parameters.
Now, lambda = 1 means that we are giving equal weighting to the data misfit and the model norm,
so this weighting gives a more intuitive feel to NLCG algorithm; the parameters should now be
independent of the relative model and data set sizes.

Also, the inversion configuration file has an extra line added for the damping parameter:
the factor by which it should be updated. Type ./Mod3DMT -I to see where.

The inversion runs, but I don't know if it's better or worse than before. All this needs to be tested!!!