Two changes.
1. 3D MT: apparent resistivities and phases. In Kristina's implementation,
one has to supply the natural log of apparent resistivities in the file.
Instead, this version takes the apparent resistivities from the file
and internally converts them to log10 for inversion; outputs apparent
resistivities (no logs). This is to avoid ambiguity in the name of the
data functional. Will need to talk to Kristina before committing this
to the stable branch.

2. Global: changes in configuration file. Here is the idea. If we need to
run the forward solver for a synthetic model that can best be expressed
in terms of a model on a grid, we use the "grid" parameterisation or
a "mixed" parameterisation which supplies a harmonic prior model and
perturbations on a grid. To accommodate this feature, we need to add
an rscalar to the model parameter in the global code.
Also, added an additional field for inversion of raw magnetic field data
(divided by something like a Dst index). This is not implemented yet,
but will be soon.