The 1D global forward solver is now reliable. It is carefully tested and cross-validated versus Jin's Matlab code (bugs in both codes fixed). All field values are correct and accurate, including for non-dipole sources.
Moreover, the Fortran code is now super-efficient: and effort has been made to only perform 
one Legenre polynomial computation for each latitude, making the minimum number of computations 
that are necessary to obtain the 1D fields.

Note: at this point, the Matlab code does not compute the vertical field at the South pole.
For P10, it is the negative of the North pole value; for non-dipole sources it can be anything,
so the quick fix currently used in Matlab is not correct. Therefore, at this point it is
preferable to use the Fortran code to obtain the 1D fields.

Will now make use of this functionality in my global SGFD code. Once implemented, will set up
several working (benchmark) examples.