Two important changes in the global.
First, I have removed the lines that increase the desired accuracy of the solution by 10 when the first period is being computed. These issues need to 
be dealt with at the user interface, or perhaps in the ForwardSolver module; certainly not within the core of the forward solver (maxwells).

Second, I have added an option to specify different forward solver parameters for the adjoint computations, and added that option to fwd_startup. If no 
such file is provided, use the solver configuration for adjoint also (default). I have fiddled with the solver configurations, and found that this is 
optimal for most problems, at 1x1 degree resolution at least:

50
1.E-6
200
30
3 3 60

These are:

ipotloopmax  ! max number of divergence correction loops
errend ! tolerance on the solution update (herr)
nrelmax    ! max number of solution updates between div. corrections
n_reldivh  ! number of divergence correction iterations
ipot0,ipotint,ipot_max ! check and run div. correction if needed every ipotmax solution updates, ipotmax = min(ipot0+ndivcorr*ipotint,ipot_max)

These parameters ensure that the divergence corrections are run often at first, when the solver is still unstable and prone to diverging, then less and 
less often at it converges. Giving the solution relaxations at least 60 iterations in between divergence corrections ensures that there is enough time 
to improve the solution (I found that previously the divergence corrections interfered with the convergence too often). Previously, the solver almost 
always reached the maximum number of iterations without convergence, and used the last 200 max relaxations to converge, without any divergence 
corrections. This resulted in a not so optimal solution and longer run times.

I have now increased the tolerance (used to be 1.E-7, in fact updated to 1.E-8 internally in the code for the first period using call_flag) so that 
this is now the primary convergence criterion. I have also increased the accuracy of divergence correction to 30 iterations (from 10): I find that this pays 
off in terms of total run time. Finally, I doubled the minimum number of solution relaxations between iterations (from 30) and increased the total 
number of divergence corrections (from 30 to 50). The solver now reaches the tolerance before the maximum number of divergence corrections is exceeded.
For now, I am making the adjoint tolerance even more relaxed, 1.E-5, but this hasn't been tested so not sure if this is appropriate.