Implemented two new data types:
T: generalised magnetic field transfer functions
H: raw magnetic fields

Sensitivities are implemented for both but haven't been tested yet.
Forward solver works with both data types.

In both cases, we read in the six components of the raw magnetic fields
at the observatories, and the errors. For T, we then scale them by a
factor which is equivalent to a Dst index, except we specify the "reference"
observatories to use in an input list. The factor is, essentially, the
average H_\theta at these reference observatories. We then invert this
response which is independent of the magnitude of the source; for output,
we "unscale" by multiplying by the same factor, saved on input.

All this needs to be tested.