I have streamlined the dependencies for the nestedEM module. Previously nestedEM used SolnSpace, and was called by EMsolve3D. It also used the modelOperator3D module.
Now, it does not use SolnSpace, and it is only called by the
modelOperator3D module. The main subroutine of the module,
Interpolate_BC_from_E_soln, has been stripped down to the basics
and moved to the ForwardSolver (could also be in the SolnSpace,
but it seems to make more sense in the ForwardSolver module).

The code changes as such have been minimal. As far as I am aware,
I haven't broken anything, but I do not have a nested modelling
example to verify this. Needs to be thourougly checked before
propagating this to the stable branch.

Of course, the way nestedEM is currently implemented, it will only work 
if there is the same number of polarizations for all transmitters.
It should be quite easy to generalise, and the new version is perhaps
easier to deal with than the old version. There is no need for the
BC_from_file array to be a vector, as far as I could see... but I haven't
changed any of that to minimize the chance of introducing errors.

[[Split portion of a mixed commit.]]