EARTH update.
I've spent the last few days debugging the Fortran global inversion... again! One would hope it'd 
be working already, after 10 years or so of development. Anyway, it seemed to behave as if the 
gradient was wrong (too big) by a factor, so I went through the code carefully and found a fresh 
bug. The inversions should converge much better now that this bug is fixed!

Briefly, in the process of making the 3D MT code more efficient (with regards to the Q business) I 
instituted a logical variable "zeroValued" in the model parameter and ensured that a bunch of 
operations, including multiplication by the model covariance, are not performed if the model 
parameter is zero valued. (Obviously, checking a full 3D model for departures from zero would not 
be efficient.)

While doing so, I also mimicked this behaviour in the global code, forgetting however to unset the 
zeroValued logical on the output from P transposed! To sum this up, ever since this bug was 
introduced the model covariance operator was never applied to the gradient, even though the model 
itself was smoothed correctly. This introduced a substantial error in the final gradient vector, 
but since all the other computations were correct, the inversion sort of worked... but not quite! 
I went though every line of NLCG and re-derived it with a pencil and paper before I found this. 

Two minor fixes still pending: need to clean up the main program Global3D.f90, and add the option
to change (increase) the NLCG sufficient decrease condition parameter c in the inverse control
configuration file.