EARTH update, mostly - and a tiny zero model output bug fix in NLCG. (The sufficient decrease condition doesn't seem to make a difference so keeping it hard coded for now).
In EARTH, I added layer weights that are user-defined but it looks like the reasonable values 
to use would be sqrt(rho), where rho is the linear prior resistivity, i.e., 10^(m0/2) if m0 
is parameterized using log10(rho). Square root since we are using CmSqrt, applied twice to form 
the model covariance Cm. This is needed to downweight the effect of inserting conductive anomalies 
in the resistive upper layers: if the data can be explained by realistic near-surface structure, 
we should accept that explanation as more probable. Relative layer weighting allows to try and 
come up with models that are heterogeneous in the uppermost layers.

Also implemented a few lines towards making the thinsheet variable, but the gradients for this are 
not implemented yet: we might not need it if we allow heterogeneity in the uppermost 100 km.

Trying it out now by running some inversion with and without the layer weighting to see if we can 
fit the data as well or better with anomalies in the upper mantle.