Global code in process of a major rewrite. Global inversion not working in this commit. It computes the first gradient, then crashes while mapping the updated model parameter to grid during initSolver. Rather mysterious, I should say. Have been debugging this (and related issues) for nearly a week and haven't found the reason yet.
The gradient works for the log10 parametrization (tested) and isn't quite right still
(by about 1/2) for arctan parametrization, which is new in the code.

Also fixed a tricky bug in the model covariance so that it now acts "correctly" (i.e.
as expected). It also means that old "unsmoothed" parametrization files shouldn't
be used with this new covariance - since it is now slightly different. The problem
was due to the fact that the "next" and "prev" coeff. pointers were not updated to
point to the respective coefficient in the new, smoothed model. So that they continued
to point to the rough model even after the vertical smoothing was run once. So the second
time around it produced an unexpected result.

Switched to using *background* model rho0 instead of the prior. This allows the "prior"
information to have arbitrary structure - not necessarily spherical harmonics - and to
try new parametrizations such as arctan. However rho0 now needs to be stored somewhere
and it's best to use a pointer so that we don't have to pass it around in MPI. That
caused another multitude of bugs until I decided to store it in modelmap. Even that
doesn't solve the problems, it seems - see above on the inversion crashing.

Most importantly, all these changes also allow varying the thinsheet layer with another
spherical harmonic layer of possibly higher degree. Once everything is debugged, will
explore this new feature.

Important output change: will now output resistivity on the grid as *.rho and unsmoothed
parametrization with all smoothing parameters as *.prm. Note that smoothed *.prm doesn't
make as much sense anymore since it is no longer the full model - background is not included.