Edited field1d.f90 to save the Legendre polynomials at grid edges and nodes, once they are computed. This is intended for efficiency, but it seems like there isn't much benefit in terms of speedup: mapping to the grid is still ~ 38 secs. Keeping the changes nevertheless.
The next edit involves saving the Legendre polynomials at grid centers and reusing them for
mapping the spherical harmonic models to the grid in modelmap.f90. I've implemented this change
and that seems to give a major speedup, reducing the time to map a degree 40 model to the grid
from ~ 55 secs to ~ 35 secs. Since we're repeating the procedure 3 times for each model
(including mapping from the 1d radial background) this might be worth keeping, but there's
currently a bug that creates NaN's. Need to fix that before I can commit.

This mapping has already been sped up by not repeating the computations for every grid layer
but for every layer in the model parametrization. It needs to be further improved: mapping from
the 1d model shouldn't take any time (currently it does, in spite of the obvious changes I made
to fix this). So, two bugs to sort out before I can leave this task. Then, need to try an MPI
inversion to verify that it still works.