Propagating Naser's fix for inter-station TFs to trunk. The fix was implemented in revs 438 and 439 in branches/stable:
Rev 438 
Small but substantial change. In the linearized data functional where Lrows for the horizontal inter-stations TF are computed, the linear 
combination of the interpolation coefficients for the x and y magnetic field components is taken between the reference-reference sites, 
which should be obviously between local-reference sites.
Bug is Fixed and the INV of inter-stations TF data is tested successfully. 

Rev 439
Slight modification to the Rx dictionary in case of inter-stations TF.
In the previous version, one field site was allowed to have only one Ref. site. In the new version one field site could be associated with 
various ref. sites.
 

[[Split portion of a mixed commit.]]