This looks like a larger number of changes, but in fact this is very minor; should work *identically* like Naser's version for 3D_MT and 3D_EM codes. This is my cleanup following Naser's major rewrite to allow for joint EM inversion. Things stopped working for me after that; mostly, the global code got broken. This is now fixed. The global code is also modified in the sense that the new elements.f90 module is included, as finalized in the semiglobal version - replacing the integral elements with finite difference approximations, to ensure exact symmetry. The global NLCG code is runs fine and produces reasonable output. The joint 3D EM code compiles, including with the PGI compiler; haven't tried running it yet. The changes are as follows.
1. Naser: in making the option case MULT_BY_J_T_multi_Tx (the -x switch), you created a fix for output of a vector of model parameters. It's in 
SensMatrix, and it is called "write_JT_multi_Tx_vec", and it in turn calls something with the name of "write_JT_multi_vec_binary" (found in 
3D_MT/modelParam/modelParamIO/Binary.inc).
Unfortunately, since SensMatrix is part of the common module package, this shortcut breaks my global code, so that it no longer compiles.

A subroutine to output a vector of model parameters was in fact already there, right next to the new one (called "writeVec_modelParam_binary"). Is 
there a reason why you thought that it was necessary to make a new version?
I have now replaced the

  call write_JT_multi_Tx_vec(JT_multi_Tx_vec,cUserDef%wFile_dModel)

in the main program Mod3DMT with the lines

  open(unit=ioSens, file=cUserDef%wFile_dModel, form='unformatted', iostat=ios)
  write(0,*) 'Output JT_multi_Tx_vec...'
  write(header,*) 'JT multi_Tx vectors'
  write(ioSens) header
  call writeVec_modelParam_binary(size(JT_multi_Tx_vec),JT_multi_Tx_vec,header,cUserDef%wFile_dModel)
  close(ioSens)

These produce identically the same output. To clean this up further, I would like to uncomment two lines in writeVec_modelParam_binary that output 1) 
the header, and 2) the size of the array. The problem is, I don't know why they were commented out, so I don't want to make any such changes before 
consulting with you, so that your output files are still readable. This function is also used to output the sensitivity matrix. For now, made sure
all output was exactly as before.

2. PGI compiler choked on some lines, which were too long: in JINV/NLCG.f90 and in 3D_EM/DC_Geoelec/DC_GeoElec.f90. I inserted line breaks to get 
everything to compile.

3. Bug fixes from Kristina Tietze, received 10 July 2014. Quoting Kristina:
"Recently, I was contacted by Juntao Cai who was intending to use inversion 
of app. res. & phase calculation with ModEM (3D). He was wondering 
about "infinity" errors for resistivities when running a forward calculation.

I had a look at the most recent version of the code (downloaded yesterday) and found that there are differences between the online version and what I 
use at two incidences:

1) DataIO: I added some lines to avoid infinity errors for resistivities (lines 194-200).

2) DataFunc: When testing the inversion of phase tensor data, I found that it is not coverging - the difference is in the Lrows-computation for phase 
tensors. There were some differences in sign convention (minus/plus; lines 553-650). I cannot remember the time when I modified that and I thought 
that I would have sent the corrected version to you, but apparently I have not - sorry for that."

These fixes need to be propagated to all other versions of the code, including the stable. Naser: I have propagated these changes to your 3D_EM 
version. The phase tensor was correct in 3D_EM, although fixed in a different way: I changed it to Kristina's code so allow fewer discrepancies 
between version. 
However, note that this also includes one very important additional fix: the derivatives for the apparent resistivities weren't previously corrected 
for the log(10) format!!!! This fix was not yet in your 3D_EM code. I have now fixed it in both versions.

Another important comment on this: in 3D_MT, the apparent resistivities in FILES are NOT log(10) values, but apparent resistivities as such; log(10) 
is taken in DataIO.f90 when the values are read in. The converse (10^value) is done on output.
This logic is missing from 3D_EM, so that the files with apparent resistivities wouldn't mean the same thing in 3D_EM and 3D_MT. This is not a good 
situation, and should probably be fixed either way - but since the 3D_MT code is already used by people, we should probably change 3D_EM. Let me know 
what you think.

4. In the global code, I updated MPI_sub.f90 to get MPI to work, and replaced elements.f90 with the new finite difference version. Since the NLCG 
works for the global code, it looks like nothing is drastically broken - although, the numbers (RMS etc) are a tiny bit different, in the second 
significant digit or so. However, something still isn't looking good in the global code because when I just tried the symmetry and gradient tests,
the gradient was nicely computed, but then the code said that it had zero norm. So something somewhere still needs fixing; but since NLCG works,
including in parallel, it most likely isn't something big.

Finally, I would like to address the issue of the multiplicity of files. I thought that we've discussed and agreed that to simplify maintenance, the 
directories 3D_EM and 3D_CSEM would only contain those files that override/supercede or add to 3D_MT, which is the base. The makefile would then 
search in these directories first, and only use 3D_MT files if the relevant files are not present in 3D_EM and 3D_CSEM. Is it ok for me to proceed to 
delete all of the duplicate files from these directories? It would make the overall structure of the code much cleaner and easier to maintain. I 
would of course compile all versions to make sure that nothing is broken. 
This release still has all the duplicates. There might be some technical issues to getting this to work as I'm envisioning; haven't tested this yet.