---------------------------------------------------------------------------------------- Changes in DataSpace module to allow for better joint inversion and related bookkeeping. ----------------------------------------------------------------------------------------
**** WARNING: MPI NOT UPDATED YET TO INCLUDE THESE CHANGES ****
**** Fixes needed in MPI_main / Master_job_Distribute_Data ****
**** Until then, cannot use enhancements for MPI inversion ****

DataSpace now includes

1) iTxt index to indicate the transmitter type: an integer that points 
   to a meaningful list in the transmitters module.
   These exist both in data blocks and in data vectors (which are single tx).

2) scalingFactor in data blocks (NOT data vectors).
   The rational is that we should keep the flexibility of separate weightings
   for each transmitter / dataType combination. In practice we will really
   want transmitter type / dataType combinations to be weighted against each other.
   The higher level modules could deal with that.

3) 
New subroutines that act on a specific transmitter type within the full dataVectorMTX:

  !**********************************************************************
  ! set the scaling of data blocks in the full data vector dataVectorMTX.
  ! there should be one for each transmitter and data type...
  ! if transmitter type iTxt is present, only affect that type of data.
  ! if iDt is present, set scaling for data blocks for this data type only.

  subroutine setScaling_dataVectorMTX(scalingFactor,d,dout,iTxt,iDt)


  !**********************************************************************
  ! extract a dataVectorMTX for specified transmitter type only.
  ! for subsetting, transmitter type index is a necessity since otherwise
  ! access to the transmitter dictionary would be required to perform
  ! this task.

  subroutine subset_dataVectorMTX(d,dout,iTxt)

Note that there is also the complementary routine,

  !**********************************************************************
  ! Merge two multi-transmitter data vectors into a single dataVectorMTX

  subroutine merge_dataVectorMTX(d1,d2,dout)

---------------------------------------------------------------------------------------
List format changes: an additional (and optional) line to determine the transmitter type.
If this line is not present, we always assume MT, so all should be backwards compatible.
The optional line is indicated by a + sign, as follows in this header example:

# Description: 
# Period(s) Code GG_Lat GG_Lon X(m) Y(m) Z(m) Component Real Imag Error
+ MT
> Full_Impedance
> exp(+i\omega t)
> [V/m]/[T]
> 0.00
> 0.000 0.000
> 4 198

COMMENT ON THE LIST FORMAT:

The header is of course meant to be completely extensible. Once you read the data type,
you are free to append as much information pertinent to that data type to the header as
you choose, and read it once for the whole block rather than repeat it in every line.
This should solve the "CSEM list format unsuitability" problem.

---------------------------------------------------------------------------------------
Important bookkeeping changes of relevance to inversion modules and DataIO.
---------------------------------------------------------------------------------------

The whole input/output module had to be reorganized to look through 
1) transmitter type index,
2) data type index,
to make it possible to, e.g., have the same data type (such as Ey_Field) defined
for multiple forcing options (e.g., CSEM and TIDE).

There is a new list - just the integer indices - of transmitter types. These are
in the transmitter dictionary module. 

  ! transmitter types; correspond to index iTxt in the data vectors
  !  these will be heavily used in inversion routines
  integer, parameter   :: MT = 1
  integer, parameter   :: DC = 2
  integer, parameter   :: CSEM = 3
  integer, parameter   :: TIDE = 4
  integer, parameter   :: GLOBAL = 5

This is all to it. In addition, there are two new subroutines tx_type_name and tx_type_index,
to toggle between the integer and character representations of transmitter type
(since key-value lists are not allowed in Fortran).
The data vector only ever has access to the integer representation. Unless you use the
transmitters.f90 module, you wouldn't know what the integers correspond to - except for
the fact that they are the same or different. I suggest that the inversion modules should
use these indices (and the transmitters.f90 module) to figure out the type of the problem,
and define the corresponding shape of the penalty functional term. The scaling factors
can be stored in the data vectors now, so the whole logic could be much streamlined.

In DataIO, the new organization required a new structure called fileInfo, which now
contains all indices and other bookkeeping details. There is one entry in the fileInfo
dictionary for each combination of transmitter type & data type encountered in the file.

---------------------------------------------------------------------------------------

TO SUM UP: 

3D_EM updated and it seems like the MPI version still works, but it will not make use of the new
	features until the MPI modules are updated as described above.
HOWEVER, STARTING FROM THIS REVISION, JOINT CODE WILL NOT WORK CORRECTLY UNLESS THE LINE 
+ CSEM
is inserted in all CSEM data blocks!

3D_MT updated for consistency and everything still works.