Propagating the air layers and LOG10 updated from semiglobal to the trunk, both 3D_EM and 3D_MT. Everything seems to work, but I have encountered an enigmatic bug that causes the gfortran version of 3D_MT to hang (in my case, txDict(1)%nPol assumed some crazy value in SolnSpace.f90)... BUT it only does that sometimes on my Mac. Then I recompile and no longer see it. Something to worry about, but it could be a compiler bug. I couldn't locate that problem in the code.
Ok, it took me a full work day but I have now fixed the air layers in Fortran. None of that has been copied to the trunk or to the stable version yet, for now it's only semiglobal/3D_EM. Everything is backwards compatible in that unless you add a couple more lines to the forward solver configuration file, the old Kush's algorithm sets the air layers by default. The new "./Mod3DMT -F" output looks like this:

Usage: -F  rFile_Model rFile_Data wFile_Data [wFile_EMsoln rFile_fwdCtrl]

 Here, rFile_fwdCtrl is the forward solver control file in the format

 Number of QMR iters per divergence correction : 40
 Maximum number of divergence correction calls : 20
 Maximum number of divergence correction iters : 100
 Misfit tolerance for EM forward solver        : 1.0e-7
 Misfit tolerance for EM adjoint solver        : 1.0e-7
 Misfit tolerance for divergence correction    : 1.0e-5
 Optional EM solution file name for nested BC  : nested.esoln

 To specify air layers, append one of these three options. Default 'mirror 10 3. 30.'

 Option 1:
 Air layers mirror|fixed height|read from file : mirror
 Number of air layers and min top dz in km     : 10 3. 30.

 Option 2:
 Air layers mirror|fixed height|read from file : fixed height
 Number of air layers and max height in km     : 12 1000.

 Option 3:
 Air layers mirror|fixed height|read from file : read from file
 Number of air layers and dz top to bottom km  : 10 500. 200. 100. 50. 20. 10. 5. 2. 1. 0.5

(if the nested file is not found, that command is ignored). I've tested against Matlab, and everything works fine. The only option which I haven't tested yet is whether or not everything still works when non-standard air layers are combined with the nested grid. The solver configuration will still work fine, but there might be some incompatibilities in interpolation from the large grid to the small.

Anyway... now that the air layers are fixed, everything pretty much works. Well, not exactly. But the problems are different. Both in Fortran and in Matlab there is about a 1% difference between the spherical and the cartesian electric fields at the surface, for the BLOCK2 model. That still seems big enough to try and track it down, but it's nothing dramatic.

At the shorter periods, the fields obtained with Matlab and with Fortran codes (either spherical or cartesian) are essentially identical, with differences of less than 0.01%. The problems start at longer periods, especially in the imaginary component: at 1000 secs, the imaginary component of the fields is almost exactly 3 times bigger in Fortran than in Matlab.

Also need to check why the setup_grid function is called so many times. We only typically need it once. There might be something that I'm missing but I have to go through this carefully to make sure. Anyway, things still work, at least in the forward code: nothing crashes, codes run and produce solutions that are very close to Matlab code. More debugging needed to understand the differences.