Change the way of modifying zt rotation, combine the rotation matrix with the calculation of Jacobin. The subroutine 'Lrows' is changed, I replaced the codes calculating partial derivatives, with the rotation matrix added. And I change some codes relative to 'Lrows', for adding a new parameter 'Azimuth'.
I deduced the equations of calculating the partial derivatives of Rotated zt with the help of Mathmatica, I think the equations are right, 
and I double checked the codes I write basing on the equations, they are the same with the equations. But, it turns out the larger the rotating angle is,
the worse the result becomes. I am still working on debugging and trying to find the reason, I just submit the codes to record these days' changes and make a checkpoint.