A few minor details changed to get unit testing to work.
1) IMPORTANT: ioAscii.f90 changed to always use integer(4) for input and output of the electric fields in binary.
This may affect other users of the code who would have by now gotten used to the fact that this depends on the
system default. However, on a Mac laptop with an Intel compiler, it did different things on a whim (integer(4)
for cartesian code, integer(8) for spherical code - not good for trying to use the files!) - Also, added some 
grid error checking on input.

2) UserCtrl.f90 changed to add fwdCtrl input file to most sensitivity tests.

3) Main_MPI.f90 changed in a minor way to add basic grid checking. Edits from stable- not pushed here yet.

4) SP/SP2 updated to reflect the latest stable-. Note however that they don't really work yet in gridelements.
Not even in cartesian coordinates. Made some crude edits to add the non-sparse grid elements to SP/SP2 to get
them to work with EMfieldInterp.f90. However, J test currently halts for some other reason. Will debug.
SP/SP2 are not at present expected to work with spherical coordinates, there's probably something else to fix
before they do. Lots, most likely.

At this point, default gridelements passes the unit testing in cartesian coords, both the symmetry tests
and the forward calculations as compared to stable- and Matlab. Note however that the L test only passes
when the numbers are large (and fails for small) so it's not very accurate for some reason. Also the impedances
differ slightly between gridelements and stable, but mostly in the 4th significant digit.

Spherical version of gridelements runs through the unit testing and supposedly passes all symmetry tests
except J test! In this latter one, J^T seems to be broken. Most specifically, it seems to be broken somewhere
after the first divergence correction. Most likely this never worked. Some major debugging to do!

I've spent a lot of time comparing the spherical and cartesian impedances, but these should not be directly
compared for the unit tests that I've set up. They are similar enough. Will also compare against Matlab.