1) BUG FIX. Most embarassing, but I missed one little but important location for the apparent resistivity and phase fixes (two releases ago). Corrected now.
2) NEW FUNCTIONALITY. This version include ModEM ADORA (with Arbitrary Data Orientation Angles) implemented by Liu Zhongyin as part of his Ph.D. project.

ModEM ADORA allows the user to specify data orientations both in bulk (through the corresponding line in the header) and line-by-line, allowing for complete
generality. At the end of each line, can add the azimuths, optionally, either for Hx, or for Hx,Hy,Ex,Ey. This allows for site-by-site orientation,
as well as (primarily for compatibility with historical data) for frequency-by-frequency orientation.
Internally, the predicted responses are rotated to correspond to the desired azimuths angles. This functionality has been tested to work in synthetic
and real inversion settings (i.e., all sensitivies are implemented for all flavors of MT responses, including the impedances, tippers, apparent resistivity
and phase, and the interstation TFs). Tilts are not currently supported, but the internal code structure supports complete generality - would be easy
to implement. Need to run this specific version in all combinations again to ensure that no bugs have been introduced.