Remove unused headers and set OBJDIR correctly
This commit removes the unnecessary, commented out header print statements at
the top of fmkmf.pl. If needed, someone can update the configuration file to
use these options, or add these options in manually.

This commit also correctly sets the OBJDIR to be what is passed in my the
configuration file via the `-o` argument.