Correct real array indexing in CSEM_module
The allocation of `bgdat % Ex`, `bgdat % Ey`, and `bgDat % Ez` was done so via
a real value. Specifying the number of elements for a variable in the allocate
statement is actually defined in the `Array elements and array sections` of the
standard (at least in the 2003/2008), there it specifies that the subscript
should be a scalar-int-expr, aka an integer, thus this isn't standard
compliant.

However, GFortran does allow for Real array indices as a legacy extension:
https://gcc.gnu.org/onlinedocs/gfortran/Real-array-indices.html.

Even more explicitly Intel specifically disallows using integer's for the
bounds:
https://www.intel.com/content/www/us/en/docs/fortran-compiler/developer-guide-reference/2024-2/allocate-statement.html

This caused GFortran to successfully compile CSEM_module.f90, with a warning of
course, but intel does not and throws an error.

I'm not sure why we are using 1.5 as a multiplier here, but we are. The fix in
this commit ensures we are passing an integer to the allocate statement.

From my testing, I believe using both `int` and the real array indices will
give the same amount of indices being allocated as `int` just truncates a real
value.