Remove included LAPACK library from ModEM
We really do not need to have the LAPACK library, mostly, because normal ModEM
does not do so and because the LAPACK and BLAS library are sometimes included
apart of certain systems.

For instance, LAPACK and BLAS are apart of the Accelerated Library on Mac.

LAPACK & BLAS use the MIT license, so we could potentially include them as a
part of ModEM. Perhaps we can do that if users' have issues compiling and
linking these libraries, but in my opinion, compiling and linking them is not
difficult. Perhaps we should provide documentation to help users compiler and
link these libraries.