Add a new Config file that works with intel and gfortran
This commit adds a new Configuration file that works with either intel or
gfortran. This is an attempt to reduce the amount of Configuration files down
to 1.

Mainly, this is an attempt to reduce the need to support or alter 10-15
configuration files if an upgrade to the config files is needed. Although this
version is on the CSEM branch, it will be a good template for something that
the main branch can use.

It requires a bit more from the user in certain situations. Specifically, it
requires users to supply the FFTW directory if they compile the EM1D csem
module. But it gives the user more flexibility, for instance they will be able
to specify which compiler flags they want (by selecting gfortran or ifort), but
they can use the FC environment variable to specify the name of their compiler
(for instance, using FC=ftn for cray machines).

The changes to this configuration file also remove the usages of `sed` in the
configuration file. Previously, the configuration file would create a
configuration file, and than alter it with `sed`. This makes the configuration
file creation confusion and harder to support. This new configuration file
builds up the `fmkmf.pl` options depending on the arguments. This should make
it easier to maintain.

To help with this transition, documentation will be added that walks users
through such options.

This configuration file also has capabilities to automatically obtain Dipole1D
if it is requested and if it is not found within `3D_MT/CSEM_module/Dipole1d`
directory.